EDRM function correction
Correct denominator of the function as the paper describes it. In the last version function, dmax was computing the highest distance possible using the prediction instead of the gold standard (true label).

Also added a boundAdjustment argument so the model output is directly clipped to the minVal and maxVal. Indeed, there are two ways to make the output clip the range, take the model output and clip it afterwards or train a model using a function with boundaries such as a sigmoid function and scale the output 0->0, 1->1/5...5->1.